requirements.txt
Flask>=2.0
pymongo>=4.0
requests>=2.28
pytest>=7.0
numpy>=1.24
python-dotenv>=1.0
groq
